Move elements to categorized folders
