+++ essentials/Icon.svelte
