+++ styled/ProgressBar.svelte
