+++ styled/ScrollTop.svelte
@@ -1,6 +1,6 @@
 <script>
 	const scrollTop = () => window.scrollTo({ top: 0, left: 0, behavior: 'smooth' });
-	import Icon from './Icon.svelte';
+	import Icon from '../essentials/Icon.svelte';
 	import { onMount } from 'svelte';
 	let y, mounted;
 	onMount(() => (mounted = true));
