+++ styled/loader/ThreeWavyBalls.svelte
